requirements.txt
fastapi
uvicorn
langdetect
pyspellchecker
spacy
errant @ git+https://github.com/chrisjbryant/errant.git
transformers